+++ src/AnyStatus.Plugins/AnyStatus.Plugins.csproj
@@ -106,15 +106,16 @@
     <Compile Include="Widgets\TeamCity\Client\TeamCityConnection.cs" />
     <Compile Include="Widgets\TeamCity\Client\TeamCityException.cs" />
     <Compile Include="Widgets\TeamCity\Build\StartTeamCityBuild.cs" />
-    <Compile Include="Widgets\TFS\BaseTfsBuildHandler.cs" />
-    <Compile Include="Widgets\TFS\OpenTfsBuildInBrowser.cs" />
-    <Compile Include="Widgets\TFS\TfsBuildMonitor.cs" />
-    <Compile Include="Widgets\TFS\TriggerTfsBuild.cs" />
+    <Compile Include="Widgets\TFS\Build\TFSBuildHandler.cs" />
+    <Compile Include="Widgets\TFS\Build\OpenWebPage.cs" />
+    <Compile Include="Widgets\TFS\Build\TFSBuildStatus.cs" />
+    <Compile Include="Widgets\TFS\Build\StartTFSBuild.cs" />
     <Compile Include="Widgets\UptimeRobot\UptimeRobotCheck.cs" />
     <Compile Include="Widgets\VSTS\Build\OpenVstsBuildPage.cs" />
     <Compile Include="Widgets\VSTS\Build\VstsBuildMonitor.cs" />
     <Compile Include="Widgets\VSTS\Build\VstsBuildTrigger.cs" />
     <Compile Include="Widgets\VSTS\Build\VstsBuild_v1.cs" />
+    <Compile Include="Widgets\VSTS\Client\VstsConnection.cs" />
     <Compile Include="Widgets\VSTS\Release\OpenVstsReleasePage.cs" />
     <Compile Include="Widgets\VSTS\Release\VSTSReleaseEnvironment.cs" />
     <Compile Include="Widgets\VSTS\Release\VSTSReleaseMonitor.cs" />
@@ -142,7 +143,7 @@
     <Compile Include="Widgets\Jenkins\View\OpenJenkinsViewInBrowser.cs" />
     <Compile Include="Widgets\Jenkins\Job\TriggerJenkinsJob.cs" />
     <Compile Include="Widgets\TeamCity\Build\TeamCityBuild.cs" />
-    <Compile Include="Widgets\TFS\TFSBuild.cs" />
+    <Compile Include="Widgets\TFS\Build\TFSBuild.cs" />
     <Compile Include="Widgets\Custom\Batch\BatchFile.cs" />
     <Compile Include="Widgets\Custom\Code\Code_v1.cs" />
     <Compile Include="Widgets\Custom\PowerShell\PowerShell.cs" />
