+++ src/AnyStatus.Plugins/Widgets/VSTS/Client/VstsConnection.cs
@@ -1,0 +1,13 @@
+﻿namespace AnyStatus
+{
+    public class VstsConnection
+    {
+        public string Account { get; set; }
+
+        public string Project { get; set; }
+
+        public string UserName { get; set; }
+
+        public string Password { get; set; }
+    }
+}
