+++ src/AnyStatus.Plugins/Widgets/VSTS/Client/VstsContracts.cs
@@ -4,16 +4,6 @@ using System.Linq;
 
 namespace AnyStatus
 {
-    public class VstsConnection
-    {
-        public string Account { get; set; }
-
-        public string Project { get; set; }
-
-        public string UserName { get; set; }
-
-        public string Password { get; set; }
-    }
 
     public class VSTSBuildDefinition
     {
